Renaming incomeSlice files, getting all refs
